Add check to MenuList component for undefined
When a child being rendered in a Menu is undefined, we need to ensure
that we don't crash. Rather, we need to simply not render the undefined
item. This is much more graceful than crashing the entire page.

The crash happens when a clone is attempted on the React element in
question, being the list item. When the child is undefined, it attempts
to clone undefined which is not a React element.