[PLT-9020] Wrap AccordionDetails in AccordionContext (#250)
* Add provideContextToChildren prop to Accordion

We want to enable AccordionDetails to be able to
access the Accordion context so that AccordionDetails
and its children can know whether or not their
Accordion is expanded. This can be used so that
we only load data when the accordion is expanded.

* Always wrap AccordionDetails in AccordionContext

There's no need to not wrap AccordionDetails in
AccordionContext.